requirements-dev.txt
bandit==1.7.5
ipykernel==6.25.2
numpy==1.26.0
pandas==2.1.1
pdfminer==20191125
PyMuPDF==1.23.4
python-generate-mac==1.3.1
requests==2.31.0
pexpect==4.8.0
